requirements.txt
langchain 
langchain-community 
langchain-core 
langchain-ollama 
chromadb 
sentence-transformers 
pypdf 
python-dotenv 
unstructured[pdf] 
tiktoken
streamlit
requests
openai
wikipedia
langchain-openai
langchain-anthropic
pydantic
duckduckgo-search

google-generativeai
